% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotate.R
\name{annotatGrWithGeneParts}
\alias{annotatGrWithGeneParts}
\title{annotatGrWithGeneParts function}
\usage{
annotatGrWithGeneParts(gr, prom, exon, intron, strand = FALSE)
}
\arguments{
\item{gr}{target object}

\item{prom}{promotors}

\item{exon}{exons}

\item{intron}{introns}

\item{strand}{logical}
}
\description{
annotatGrWithGeneParts function
}
