% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotate.R
\name{as.data.frame,AnnotationByGeneParts-method}
\alias{as.data.frame,AnnotationByGeneParts-method}
\title{Convert annotated object to a data.frame}
\usage{
\S4method{as.data.frame}{AnnotationByGeneParts}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{AnnotationByGeneParts}.}

\item{row.names, optional, ...}{additional arguments to be passed to or from methods.}
}
\value{
a data.frame
}
\description{
Convert annotated object from \code{annotateWithGeneParts()} to a data.frame,
merging the `members` and `dist.to.TSS` slots.
}
\examples{
data(cage)
bed.file = system.file("extdata/chr21.refseq.hg19.bed", package = "genomation")
gene.parts = readTranscriptFeatures(bed.file)
cage.annot = annotateWithGeneParts(cage, gene.parts, intersect.chr=TRUE)
as.data.frame(cage.annot)

}
