% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotate.R
\name{bed12ToExons}
\alias{bed12ToExons}
\title{bed12ToExons function}
\usage{
bed12ToExons(ref)
}
\arguments{
\item{ref}{data.frame object}
}
\description{
extracts exons from a bed12 file and puts them into GRanges object
}
\keyword{internal}
