% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotate.R
\name{bed12ToIntrons}
\alias{bed12ToIntrons}
\title{bed12ToIntrons function}
\usage{
bed12ToIntrons(ref)
}
\arguments{
\item{ref}{data.frame object}
}
\description{
extracts introns from a bed12 file and puts them into GRanges object
}
