% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binMean}
\alias{binMean}
\alias{binMedian}
\alias{binMax}
\alias{binMin}
\alias{binSum}
\title{Functions that compute the value for each bin}
\usage{
binMean(x, n)

binMedian(x, n)

binMax(x, n)

binMin(x, n)

binSum(x, n)
}
\arguments{
\item{x}{NumericVector - vector of values of a bin}

\item{n}{intiger - number of bins}
}
\description{
binMean() computes a mean value, binMedian() computes a median value, binMax() and binMin() 
give maximum and minumum values respectively, binSum() computes the sum of the values in a bin
}
