% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binMean}
\alias{binMean}
\title{Function that computes a mean value for each bin}
\usage{
binMean(x, n)
}
\arguments{
\item{x}{NumericVector}

\item{n}{intiger - number of bins}
}
\description{
Function that computes a mean value for each bin
}
\keyword{internal}
