% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binMedian}
\alias{binMedian}
\title{Function that computes a median value for each bin}
\usage{
binMedian(x, n)
}
\arguments{
\item{x}{NumericVector - vector of values of a bin}

\item{n}{intiger - number of bins}
}
\description{
Function that computes a median value for each bin
}
\keyword{internal}
