% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binMin}
\alias{binMin}
\title{Function that computes a minimum value for each bin}
\usage{
binMin(x, n)
}
\arguments{
\item{x}{NumericVector - vector of values of a bin}

\item{n}{intiger - number of bins}
}
\description{
Function that computes a minimum value for each bin
}
\keyword{internal}
