% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{binner}
\alias{binner}
\title{binner function}
\usage{
binner(start, end, nbins)
}
\arguments{
\item{start}{start position}

\item{end}{end position}

\item{nbins}{number of bins}
}
\description{
given a vector and length smooths the vector to a given size
}
