% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binSum}
\alias{binSum}
\title{Function that computes a sum of values in a bin}
\usage{
binSum(x, n)
}
\arguments{
\item{x}{NumericVector - vector of values of a bin}

\item{n}{intiger - number of bins}
}
\description{
Function that computes a sum of values in a bin
}
\keyword{internal}
