% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineScoreMatrixList.R
\name{c.ScoreMatrix}
\alias{c.ScoreMatrix}
\title{c.ScoreMatrix}
\usage{
\method{c}{ScoreMatrix}(..., recursive = FALSE, use.names = TRUE)
}
\arguments{
\item{...}{contais scoreMatrix and scoreMatrixList objects}

\item{recursive}{logical}

\item{use.names}{logical}
}
\value{
returns a \code{scoreMatrixList} object
}
\description{
Combine a scoreMatrix into a scoreMatrixList object - when a ScoreMatrix is a first argument
}
