% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineScoreMatrixList.R
\name{c.ScoreMatrixList}
\alias{c.ScoreMatrixList}
\title{c.ScoreMatrixList}
\usage{
\method{c}{ScoreMatrixList}(..., recursive = FALSE, use.names = TRUE)
}
\arguments{
\item{...}{contais scoreMatrix and scoreMatrixList objects}

\item{recursive}{logical}

\item{use.names}{logical}
}
\value{
returns a \code{scoreMatrixList} object
}
\description{
Combine a scoreMatrix into a scoreMatrixList object - when a ScoreMatrixList is a first argument
}
