% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotate.R
\name{checkBedValidity}
\alias{checkBedValidity}
\title{checkBedValidity function}
\usage{
checkBedValidity(bed.df, type = "none")
}
\arguments{
\item{bed.df}{data.frame object}

\item{type}{type}
}
\description{
checks the validity of the bed data.frame if it is a legitimate bed columns
}
