% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{checkClass}
\alias{checkClass}
\title{checkClass function}
\usage{
checkClass(x, class.name, var.name = deparse(substitute(x)))
}
\arguments{
\item{x}{object}

\item{class.name}{class name}

\item{var.name}{uses x object}
}
\description{
check whether the x object corresponds to the given class
}
\keyword{internal}
