% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{compressedAndUrl2temp}
\alias{compressedAndUrl2temp}
\title{compressedAndUrl2temp function}
\usage{
compressedAndUrl2temp(filename)
}
\arguments{
\item{filename}{file name}
}
\description{
compressedAndUrl2temp function
}
