% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{constrainRanges}
\alias{constrainRanges}
\title{constrainRanges function}
\usage{
constrainRanges(target, windows)
}
\arguments{
\item{target}{target file}

\item{windows}{windows}
}
\description{
removes ranges that fell of the rle object
does not check for the correspondence of the chromosome names - always 
check before using this function
}
