% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{detectUCSCheader}
\alias{detectUCSCheader}
\title{detectUCSCheader function}
\usage{
detectUCSCheader(filename)
}
\arguments{
\item{filename}{file name}
}
\description{
detects UCSC header (and first track)
}
