% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotate.R
\name{distance2NearestFeature}
\alias{distance2NearestFeature}
\title{distance2NearestFeature function}
\usage{
distance2NearestFeature(g.idh, tss)
}
\arguments{
\item{g.idh}{target object}

\item{tss}{TSSes}
}
\description{
distance2NearestFeature function
}
