% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentMatrix.R
\name{enrichmentMatrix}
\alias{enrichmentMatrix}
\alias{enrichmentMatrix,ScoreMatrixControl-method}
\title{Compute an enrichment of IP over control both stored in ScoreMatrixControl object}
\usage{
enrichmentMatrix(x)
}
\arguments{
\item{x}{\code{\link{ScoreMatrixControl}}}
}
\value{
\code{\link{ScoreMatrix}}
}
\description{
This is a \code{enrichmentMatrix} function for \code{ScoreMatrixControl} object, 
that enables to normalize ChIP-seq signals with respect to IgG or input DNA control.
}
\examples{
 
#load IP and control BAM files and create ScoreMatrix objects
source("http://bioconductor.org/biocLite.R")
biocLite('genomationData')
bam.file_IP <- system.file("extdata", 
"wgEncodeBroadHistoneH1hescSuz12051317AlnRep1.chr21.bam", package = "genomationData")
bam.file_c <- system.file("extdata", 
"wgEncodeBroadHistoneH1hescControlStdAlnRep1.chr21.bam", package = "genomation")
data(promoters)
IP <- ScoreMatrix(target = bam.file_IP, windows = promoters, type = 'bam')
control <- ScoreMatrix(target = bam.file_c, windows = promoters, type = 'bam')

# create a ScoreMatrixControl object
smControl <- ScoreMatrixControl(IP, control)

# compute an enrichment of IP over control
enrichmentMatrix(smControl)

}
\seealso{
\code{\link{ScoreMatrixListControl}}, \code{\link{ScoreMatrixControl}}
}
