% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentMatrix.R
\docType{methods}
\name{enrichmentMatrix,ScoreMatrixListControl-method}
\alias{enrichmentMatrix,ScoreMatrixListControl-method}
\title{Compute an enrichment of IP over control both stored in ScoreMatrixListControl object}
\usage{
\S4method{enrichmentMatrix}{ScoreMatrixListControl}(x)
}
\arguments{
\item{x}{\code{\link{ScoreMatrixListControl}}}
}
\value{
\code{\link{ScoreMatrixList}}
}
\description{
This is a \code{\link{enrichmentMatrix}} function for ScoreMatrixListControl object, that enables to normalize ChIP-seq signals with respect to IgG or input DNA control.
}
