% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{genes}
\alias{genes}
\title{Example RefSeq genes data set.}
\format{\code{\link{GRanges}} object}
\description{
RefSeq genes of hg19 assembly of human genome on chr21 and chr22. Downloaded
from UCSC genome browser.
}

