% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getMembers}
\alias{getMembers}
\alias{getMembers,AnnotationByFeature-method}
\title{Get the membership slot of AnnotationByFeature}
\usage{
getMembers(x)

\S4method{getMembers}{AnnotationByFeature}(x)
}
\arguments{
\item{x}{a \code{AnnotationByFeature}  object}
}
\value{
matrix showing overlap of target features with annotation features.
        1 for overlap, 0 for non-overlap
}
\description{
Membership slot defines the overlap of target features with annotation features
 For example, if a target feature overlaps with an exon
}

