\docType{methods}
\name{heatmapProfile}
\alias{heatmapProfile}
\alias{heatmapProfile,scoreMatrixList-method}
\title{Plot a scoreMatrixList object, along with the underlying factors}
\usage{
  heatmapProfile(mat.list, mat.cols=NULL, ...)
}
\arguments{
  \item{mat.list}{a scoreMatrixList object}

  \item{mat.cols}{colors to be used for plotting}

  \item{...}{other options (obselete for now)}
}
\value{
  returns a \code{scoreMatrixList} object
}
\description{
  Plot a scoreMatrixList object, along with the underlying
  factors
}

