% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idxStats}
\alias{idxStats}
\title{Fast retrieval of indexstats}
\usage{
idxStats(bam_file)
}
\arguments{
\item{bam_file}{path to the indexed BAM file}
}
\description{
Retrieve and print stats in the index file corresponding 
to the input file. Before calling idxstats, the input BAM file
must be indexed, e.g. via samtools index.
}

