% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listSliceMean}
\alias{listSliceMean}
\alias{listSliceMedian}
\alias{listSliceMax}
\alias{listSliceMin}
\alias{listSliceSum}
\title{Functions create a matrix storing the data with desirable number of bins for each window}
\usage{
listSliceMean(xlist, n, ranks)

listSliceMedian(xlist, n, ranks)

listSliceMax(xlist, n, ranks)

listSliceMin(xlist, n, ranks)

listSliceSum(xlist, n, ranks)
}
\arguments{
\item{xlist}{List of vectors storing values of a bin}

\item{n}{intiger - number of bins}

\item{ranks}{CharacterVector - position of the windows whose strand is "-"}
}
\description{
listSliceMean() calls the binMean() function, listSliceMedian calls the binMedian(), 
listSliceMax() - binMax(), listSliceMin() - binMin(), listSliceSum() - binSum()
}
