% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listSliceMax}
\alias{listSliceMax}
\title{Function creates a matrix storing data with desirable number of bins for each window}
\usage{
listSliceMax(xlist, n)
}
\arguments{
\item{xlist}{List of vectors storing values of a bin}

\item{n}{intiger - number of bins}
}
\description{
listSliceMax() function calls the binMax() function
}
\keyword{internal}
