% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listSliceMean}
\alias{listSliceMean}
\title{Function creates a matrix storing data with desirable number of bins for each window}
\usage{
listSliceMean(xlist, n)
}
\arguments{
\item{xlist}{List of vectors storing values of a bin}

\item{n}{intiger - number of bins}
}
\description{
listSliceMean() function calls the binMean() function
}
\keyword{internal}
