% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listSliceMin}
\alias{listSliceMin}
\title{Function creates a matrix storing data with desirable number of bins for each window}
\usage{
listSliceMin(xlist, n)
}
\arguments{
\item{xlist}{List of vectors storing values of a bin}

\item{n}{intiger - number of bins}
}
\description{
listSliceMin() function calls the binMin() function
}
\keyword{internal}
