% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{listSliceSum}
\alias{listSliceSum}
\title{Function creates a matrix storing data with desirable number of bins for each window}
\usage{
listSliceSum(xlist, n)
}
\arguments{
\item{xlist}{List of vectors storing values of a bin}

\item{n}{intiger - number of bins}
}
\description{
listSliceSum() function calls the binSum() function
}
\keyword{internal}
