% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matRes}
\alias{matRes}
\title{Function computes a matrix that stores the data with desirable number of bins and 
keeps original order of the windows}
\usage{
matRes(xlist, n, negranks, binOp, p)
}
\arguments{
\item{xlist}{List of vectors storing values of a bin}

\item{n}{intiger - number of bins}

\item{negranks}{CharacterVector - position of the windows whose strand is "-"}

\item{p}{NumericVector - stors an original window order}
}
\description{
Function computes a matrix that stores the data with desirable number of bins and 
keeps original order of the windows
}
