% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Median_c}
\alias{Median_c}
\title{Function that computes a median value}
\usage{
Median_c(x)
}
\arguments{
\item{x}{NumericVector}
}
\description{
Function that computes a median value
}
\keyword{internal}
