% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{Ops,ScoreMatrix,ScoreMatrix-method}
\alias{Ops,ScoreMatrix,ScoreMatrix-method}
\title{Ops method for a ScoreMatrix object. It enables to use arithmetic, indicator and logic operations on ScoreMatrix objects.}
\usage{
\S4method{Ops}{ScoreMatrix,ScoreMatrix}(e1, e2)
}
\arguments{
\item{e1}{the \code{\link{ScoreMatrix}} object or numeric value}

\item{e2}{the \code{\link{ScoreMatrix}} object or numeric value}
}
\value{
\code{ScoreMatrix}
}
\description{
Arithmetic method for ScoreMatrix and ScoreMatrixList classes
}

