% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{Ops,ScoreMatrixList,ScoreMatrixList-method}
\alias{Ops,ScoreMatrixList,ScoreMatrixList-method}
\title{Ops method for a ScoreMatrixList object. It enables to use arithmetic, indicator and logic operations on ScoreMatrixList objects.}
\usage{
\S4method{Ops}{ScoreMatrixList,ScoreMatrixList}(e1, e2)
}
\arguments{
\item{e1}{the \code{\link{ScoreMatrixList}} object}

\item{e2}{the \code{\link{ScoreMatrixList}} object}
}
\value{
\code{ScoreMatrixList}
}
\description{
Arithmetic methods for ScoreMatrixList
}

