\docType{methods}
\name{order,ScoreMatrixList-method}
\alias{order,ScoreMatrixList-method}
\title{Reorder all elements of a ScoreMatrixList to a given ordering vector}
\usage{
  order(sml, ord.vec)
}
\arguments{
  \item{sml}{\code{ScoreMatrixList} object}

  \item{ord.vec}{an integer vector}
}
\value{
  \code{ScoreMatrixList} object
}
\description{
  Reorder all elements of a ScoreMatrixList to a given
  ordering vector
}

