\docType{methods}
\name{order}
\alias{order}
\title{Reorder all elements of a ScoreMatrixList to a given ordering vector}
\arguments{
  \item{sml}{\code{ScoreMatrixList} object}

  \item{ord.vec}{an integer vector}
}
\value{
  \code{ScoreMatrixList} object
}
\description{
  Reorder all elements of a ScoreMatrixList to a given
  ordering vector
}

