\docType{methods}
\name{orderBy}
\alias{orderBy}
\alias{orderBy,ScoreMatrixList-method}
\title{Reorder all elements of a ScoreMatrixList to a given ordering vector}
\usage{
  orderBy(sml, ord.vec)

  \S4method{orderBy}{ScoreMatrixList}(sml, ord.vec)
}
\arguments{
  \item{sml}{\code{ScoreMatrixList} object}

  \item{ord.vec}{an integer vector}
}
\value{
  \code{ScoreMatrixList} object
}
\description{
  Reorder all elements of a ScoreMatrixList to a given
  ordering vector
}

