% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patternMatrix.R
\docType{methods}
\name{patternMatrix}
\alias{patternMatrix}
\title{Get scores that correspond to k-mer occurence for bases in each window}
\usage{
patternMatrix(pattern, windows, genome = NULL, min.score = NULL,
  cores = 1, asPercentage = FALSE, occurence = TRUE)
}
\arguments{
\item{pattern}{matrix, a PWM matrix. It has one row for each nucleotide ("A","C","G" and "T")}

\item{windows}{\code{GRanges} object}

\item{genome}{\code{BSgenome} object}

\item{min.score}{numeric or string indicating minimum score for counting a match.
Can be given as a character string containing
a percentage of the highest possible score or as a single number (as default "80%").}

\item{cores}{the number of cores to use (default: 1)}

\item{asPercentage}{boolean telling whether scores represent percentage of the maximal
motif PWM score (default: TRUE) or raw scores (FALSE).}

\item{occurence}{if pattern is a matrix then function can either return occurence
(a scoreMatrix with 1s and 0s for k-mers)
using given score cutoff \code{min.score} (default: TRUE)
or scores themselves (FALSE)}
}
\value{
returns a \code{scoreMatrix} object
}
\description{
we are looking for occurrence, we aim to get a scoreMatrix with 1s and 0s for k-mers.
The function first look for and calculates
The function firsts bins each window to equal number of bins, and calculates
the a summary matrix for scores of each bin (currently, mean, max and min supported)
A patternMatrix object can be used to ..
\code{windows} can be a predefined region such as CpG islands or gene bodies that have the same width.
}
\details{
we are looking for occurrence, we aim to get a scoreMatrix with 1s and 0s for k-mers.
}
\seealso{
\code{\link{ScoreMatrix}}
}

