% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{promoters}
\alias{promoters}
\title{Example promoter data set.}
\format{\code{\link{GRanges}} object}
\description{
promoters of hg19 assembly of human genome on chr21 and chr22. Promoter set is 
derived from refseq TSS.
}
