% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ranksOrder}
\alias{ranksOrder}
\title{Function reorders matrix to obtain the original order of the windows}
\usage{
ranksOrder(x, p)
}
\arguments{
\item{x}{NumericMatrix}

\item{p}{NumericVector - stors an original window order}
}
\description{
Function reorders matrix to obtain the original order of the windows
}
