% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{readBam}
\alias{readBam}
\title{readBam function}
\usage{
readBam(target, windows, rpm = FALSE, unique = FALSE, extend = 0,
  param = NULL, paired.end = FALSE, library.size = NULL, ...)
}
\arguments{
\item{target}{target object}

\item{windows}{windows}

\item{rpm}{logical}

\item{unique}{logical}

\item{extend}{numeric}

\item{param}{ScanBamParam object}

\item{paired.end}{llogical}

\item{library.size}{numeric}

\item{...}{additional parameters}
}
\description{
given a big bam path reads the big wig file into a RleList
to be used by ScoreMatrix:char,GRanges
}
\keyword{internal}
