% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{readBigWig}
\alias{readBigWig}
\title{readBigWig function}
\usage{
readBigWig(target, windows = NULL, ...)
}
\arguments{
\item{target}{target object}

\item{windows}{windows}

\item{...}{additional parameters}
}
\description{
given a big wig path reads the big wig file into a RleList
to be used by ScoreMatrix:char,GRanges
}
