% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readTableFast}
\alias{readTableFast}
\title{readTableFast function}
\usage{
readTableFast(filename, header = TRUE, skip = 0, sep = "\\t", chr = 1)
}
\arguments{
\item{filename}{file name}

\item{header}{logical}

\item{skip}{numeric}

\item{sep}{character}

\item{chr}{numeric}
}
\description{
fast reading of big tables
chr indicates index of column of chromosomes
}
\keyword{internal}
