\name{scoreMatrix-class}
\alias{scoreMatrix-class}
\title{An S4 class for storing \code{scoreMatrix} function results}
\description{
  The resulting object is an extension of a \code{matrix}
  object, and stores values (typically genome-wide scores)
  for a predefined set of regions Each row on the
  scoreMatrix is a predefined region (Ex: CpG islands,
  promoters) and columns are values across those regions.
}
\seealso{
  \code{\link{scoreMatrix-methods}}
}

