% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomation-classes.R
\docType{class}
\name{ScoreMatrixControl-class}
\alias{ScoreMatrixControl-class}
\title{An S4 class for storing \code{ScoreMatrixControl} function results}
\description{
The resulting object is an extension of a \code{ScoreMatrix} object, that stores ScoreMatrix of IP sample as well as ScoreMatrix containing IgG or input DNA control sample.
}
\section{Slots}{

\describe{
\item{\code{control}}{A \code{\link{ScoreMatrix}} object storing a control sample.}
}}

\section{Constructors}{

see \code{\link{ScoreMatrixControl}}
}

\section{Coercion}{

as(from, "ScoreMatrixControl"): Creates a \code{ScoreMatrixControl} object from a \code{\link{ScoreMatrix}} or \code{\link{ScoreMatrixBin}} objects.
}

\section{Subsetting}{

In the code snippets below, x is a ScoreMatrixControl object.
 
\code{'x[i,j]'}: Get or set elements from row \code{i} and column \code{j} and return a subset of IP ScoreMatrix object.

\code{'x@control[i,j]'}: Get or set elements from row \code{i} and column \code{j} and return a subset of control ScoreMatrix object.
}

\seealso{
\code{\link{ScoreMatrixListControl}}, \code{\link{ScoreMatrix}}
}
