% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrixControl.R
\docType{methods}
\name{ScoreMatrixControl}
\alias{ScoreMatrixControl}
\alias{ScoreMatrixControl,ScoreMatrix,ScoreMatrix}
\alias{-}
\alias{method}
\title{Make ScoreMatrixControl from ScoreMatrix objects of IP and control samples}
\usage{
ScoreMatrixControl(IP, control)
}
\arguments{
\item{IP}{the \code{\link{ScoreMatrix}} object storing an IP sample}

\item{control}{the \code{\link{ScoreMatrix}} object storing an control sample}
}
\value{
returns a \code{ScoreMatrixControl} object
}
\description{
The function creates a \code{ScoreMatrixControl} object that stores a \code{ScoreMatrix} containing an signal values of IP sample as well as \code{ScoreMatrix} object with control sample. 
This object is an input in \code{\link{enrichmentMatrix}} function that compute enrichment over IgG or input DNA control.
}
