% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomation-classes.R
\docType{class}
\name{ScoreMatrixListControl-class}
\alias{ScoreMatrixListControl-class}
\title{An S4 class for storing \code{ScoreMatrixListControl} function results}
\description{
The resulting object is an extension of a \code{ScoreMatrixList} object, that stores \code{ScoreMatrixControl} objects
}
\section{Slots}{

\describe{
\item{\code{names}}{A character vector containing names of all stored \code{ScoreMatrixControl} objects.}
}}

\section{Constructors}{

see \code{\link{ScoreMatrixListControl}}
}

\section{Coercion}{

as(from, "ScoreMatrixListControl"): Creates a \code{ScoreMatrixListControl} object from a list of \code{\link{ScoreMatrixControl}} objects.
}

\section{Subsetting}{

In the code snippets below, x is a ScoreMatrixListControl object.
 
\code{x[[i]]}: Get or set \code{ScoreMatrixControl} object \code{i}, where \code{i} is a numeric or character vector of length 1.

\code{x$name}: value: Get or set element \code{name}, where \code{name} is a name or character vector of length 1.
}

\seealso{
\code{\link{ScoreMatrixControl}}, \code{\link{ScoreMatrixList}}
}
