% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrixListControl.R
\docType{methods}
\name{ScoreMatrixListControl,ScoreMatrixList,ScoreMatrix-method}
\alias{ScoreMatrixListControl,ScoreMatrixList,ScoreMatrix-method}
\title{Make ScoreMatrixListControl from ScoreMatrixList object of IP samples and ScoreMatrix object of control sample}
\usage{
\S4method{ScoreMatrixListControl}{ScoreMatrixList,ScoreMatrix}(IP, control)
}
\arguments{
\item{IP}{\code{\link{ScoreMatrixList}} object storing an IP samples}

\item{control}{\code{\link{ScoreMatrix}} storing control sample}
}
\value{
\code{ScoreMatrixListControl} object
}
\description{
The function creates a \code{ScoreMatrixListControl} object that stores \code{ScoreMatrixControl} objects.
This object is an input in \code{\link{enrichmentMatrix}} function that compute enrichment over IgG or input DNA control.
}
