% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomation-classes.R, R/readAnnotate.R,
%   R/scoreMatrix.R, R/scoreMatrixControl.R, R/scoreMatrixList.R,
%   R/scoreMatrixListControl.R
\docType{methods}
\name{show,RandomEnrichment-method}
\alias{show,RandomEnrichment-method}
\alias{show,AnnotationByGeneParts-method}
\alias{show,AnnotationByFeature-method}
\alias{show,ScoreMatrix-method}
\alias{show,ScoreMatrixControl-method}
\alias{show,ScoreMatrixList-method}
\alias{show,ScoreMatrixListControl-method}
\title{show method for some of the genomation classes}
\usage{
\S4method{show}{RandomEnrichment}(object)

\S4method{show}{AnnotationByGeneParts}(object)

\S4method{show}{AnnotationByFeature}(object)

\S4method{show}{ScoreMatrix}(object)

\S4method{show}{ScoreMatrixControl}(object)

\S4method{show}{ScoreMatrixList}(object)

\S4method{show}{ScoreMatrixListControl}(object)
}
\arguments{
\item{object}{object of class RandomEnrichment}
}
\value{
Shows the dimension of the ScoreMatrix

Shows the sizes of both IP and control matrices stored in a ScoreMatrixControl object

Shows the number of matrices and their sizes

Shows the number of stored ScoreMatrixControl objects and their sizes
}
\description{
show method for some of the genomation classes
}
