% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{[,ScoreMatrix,ANY,ANY-method}
\alias{Ops,ScoreMatrix-method}
\alias{[,ScoreMatrix,ANY,ANY-method}
\title{Subsetting method for a ScoreMatrix object.}
\usage{
\S4method{[}{ScoreMatrix,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{the \code{\link{ScoreMatrix}} object}

\item{i}{numeric value}

\item{j}{numeric value}
}
\value{
\code{ScoreMatrix}
}
\description{
Subsetting method for a ScoreMatrix object.
}

