% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{[,ScoreMatrixList,ANY,ANY,ANY-method}
\alias{[,ScoreMatrixList,ANY,ANY,ANY-method}
\alias{[,ScoreMatrixList-method}
\title{Extract method for a ScoreMatrixList object.}
\usage{
\S4method{[}{ScoreMatrixList,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{the \code{\link{ScoreMatrixList}} object}

\item{i}{numeric value}
}
\description{
Extract method for a ScoreMatrixList object.
}
