% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{[,ScoreMatrixList,ANY,ANY-method}
\alias{Ops,ScoreMatrixList-method}
\alias{[,ScoreMatrixList,ANY,ANY-method}
\title{Subsetting method for a ScoreMatrixList object}
\usage{
\S4method{[}{ScoreMatrixList,ANY,ANY}(x, i)
}
\arguments{
\item{x}{the \code{\link{ScoreMatrixList}} object}

\item{i}{numeric value}
}
\value{
\code{ScoreMatrixList}
}
\description{
Subsetting method for a ScoreMatrixList object
}

