% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrix.R
\name{target.type}
\alias{target.type}
\title{target.type function}
\usage{
target.type(target, type = "")
}
\arguments{
\item{target}{target file}

\item{type}{type name}
}
\description{
Check if a target file is in bam and bigWig formats
by looking at the file extension
}
