% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{cage}
\alias{cage}
\title{Example CAGE data set.}
\format{\code{\link{GRanges}} object}
\description{
Location and tag per million values for CAGE TSS clusters on chr21 and chr22
of human genome (hg19 assembly). The clusters are dowloaded from ENCODE project
downloads for NHEK cells.
}
