% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{cpgi}
\alias{cpgi}
\title{Example CpG island data set.}
\format{\code{\link{GRanges}} object}
\description{
CpG islands of hg19 assembly of human genome on chr21 and chr22. Downloaded
from UCSC genome browser.
}

