% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentMatrix.R
\name{enrichmentMatrix}
\alias{enrichmentMatrix}
\alias{enrichmentMatrix,ScoreMatrixControl-method}
\title{Compute an enrichment of IP over control stored in ScoreMatrixControl object}
\usage{
enrichmentMatrix(x)
}
\arguments{
\item{x}{the \code{\link{ScoreMatrixControl}} object}
}
\value{
\code{ScoreMatrix}
}
\description{
This is a \code{\link{enrichmentMatrix}} function for a ScoreMatrixControl object that enables to normalize ChIP-seq signals with respect to IgG or input DNA control.
}
