% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentMatrix.R
\docType{methods}
\name{enrichmentMatrix,ScoreMatrixListControl-method}
\alias{enrichmentMatrix,ScoreMatrixListControl-method}
\title{Compute an enrichment of IP over control both stored in ScoreMatrixListControl object}
\usage{
\S4method{enrichmentMatrix}{ScoreMatrixListControl}(x)
}
\arguments{
\item{x}{\code{\link{ScoreMatrixListControl}}}
}
\value{
\code{\link{ScoreMatrixList}}
}
\description{
This is a \code{enrichmentMatrix} function for \code{ScoreMatrixListControl} object, 
that enables to normalize ChIP-seq signals with respect to IgG or input DNA control.
}
\examples{
#load IP and control BAM files and create ScoreMatrix objects
source("http://bioconductor.org/biocLite.R")
biocLite('genomationData')
data(promoters)
bam.file_IP_1 <- system.file("extdata", 
"wgEncodeSydhTfbsH1hescZnf143IggrabAlnRep1.chr21.bam", package = "genomationData")
IP_1 <- ScoreMatrix(target = bam.file_IP_1, windows = promoters, type = 'bam')

bam.file_IP_2 <- system.file("extdata", 
"wgEncodeBroadHistoneH1hescSuz12051317AlnRep1.chr21.bam", package = "genomationData")
IP_2 <- ScoreMatrix(target=bam.file_IP_2, windows = promoters, type = 'bam')

bam.file_c <- system.file("extdata", 
"wgEncodeBroadHistoneH1hescControlStdAlnRep1.chr21.bam", package = "genomation")
control <- ScoreMatrix(target = bam.file_c, windows = promoters, type = 'bam')

# create a ScoreMatrixList object of IP ScoreMatrix object
IPs <- list(IP_1,IP_2)
sml_IP <- ScoreMatrixList(IPs)

#create a ScoreMatrixListControl object
smlControl <- ScoreMatrixListControl(sml_IP, control)

# compute an enrichment of IP over control
enrichmentMatrix(smlControl)

}
