% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{Ops,numeric,ScoreMatrixList-method}
\alias{Ops,ScoreMatrixList-method}
\alias{Ops,numeric,ScoreMatrixList-method}
\title{Ops method for a ScoreMatrixList object. It enables to use arithmetic, indicator and logic operations on ScoreMatrixList objects.}
\usage{
\S4method{Ops}{numeric,ScoreMatrixList}(e1, e2)
}
\arguments{
\item{e1}{the numeric value}

\item{e2}{the \code{\link{ScoreMatrixList}} object}
}
\value{
\code{ScoreMatrixList}
}
\description{
Ops method for a ScoreMatrixList object. It enables to use arithmetic, indicator and logic operations on ScoreMatrixList objects.
}

