% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readAnnotate.R
\docType{methods}
\name{plotGeneAnnotation}
\alias{plotGeneAnnotation}
\alias{plotGeneAnnotation,AnnotationByFeature-method}
\alias{plotGeneAnnotation,list-method}
\title{Plots the enrichment of each feature in the set in the gene annotation}
\usage{
plotGeneAnnotation(l, cluster = FALSE, col = c("white", "cornflowerblue"))

\S4method{plotGeneAnnotation}{list}(l, cluster = FALSE, col = c("white",
  "cornflowerblue"))
}
\arguments{
\item{l}{a \code{list} of  \code{AnnotationByGeneParts} objects}

\item{cluster}{TRUE/FALSE. If TRUE the heatmap is going to be clustered
using hierarchical clustering}

\item{col}{a vector of two colors that will be used for interpolation.
       The first color is the lowest one, the second is the highest one}
}
\value{
plots a heatmap of enrichment of target in each gene functional
        category
}
\description{
This function plots a heatmap of enrichment of each range in given gene feature
}
\examples{
# library(GenomicRanges)
# data(cage)
# data(cpgi)

# cage$tpm = NULL
# gl = GRangesList(cage=cage, cpgi=cpgi)

# bed.file = system.file("extdata/chr21.refseq.hg19.bed", package = "genomation")
# gene.parts = readTranscriptFeatures(bed.file)
# annot = annotateWithGeneParts(gl, gene.parts, intersect.chr=TRUE)
# plotGeneAnnotation(annot)
}

