% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readData.R
\docType{methods}
\name{readBroadPeak}
\alias{readBroadPeak}
\title{A function to read the Encode formatted broad peak file into a GRanges object}
\usage{
readBroadPeak(file)
}
\arguments{
\item{file}{a abosulte or relative path to a bed file formatted by the Encode broadPeak standard}
}
\value{
a GRanges object
}
\description{
A function to read the Encode formatted broad peak file into a GRanges object
}
\examples{
# broad.peak.file = system.file('extdata',"ex.broadPeak", package='genomation')

# broad.peak = readBroadPeak(broad.peak.file)
# head(broad.peak)
}

