% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/readData.R
\docType{methods}
\name{readBroadPeak}
\alias{readBroadPeak}
\title{A function to read the Encode formatted broad peak file into a GRanges object}
\usage{
readBroadPeak(file, track.line=FALSE)
}
\arguments{
\item{file}{an absolute or relative path to a bed file formatted by the Encode broadPeak standard.
The file can end in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip}
and/or start with \code{http://} or \code{ftp://}. If the file is not compressed
it can also start with \code{https://} or \code{ftps://}.}

\item{track.line}{the number of track lines to skip, "auto" to detect them automatically
or FALSE(default) if the bed file doesn't have track lines}
}
\value{
a GRanges object
}
\description{
A function to read the Encode formatted broad peak file into a GRanges object
}
\examples{
# broad.peak.file = system.file('extdata',"ex.broadPeak", package='genomation')

# broad.peak = readBroadPeak(broad.peak.file)
# head(broad.peak)
}

