% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ScoreMatrix-class}
\alias{ScoreMatrix-class}
\title{An S4 class for storing \code{ScoreMatrix} function results}
\description{
The resulting object is an extension of a \code{matrix} object, and stores values (typically genome-wide scores) for a predefined set of regions
Each row on the ScoreMatrix is a predefined region (Ex: CpG islands, promoters) and columns are values across those regions.
}
\seealso{
\code{\link{ScoreMatrix}}
}

