% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrixControl.R
\docType{methods}
\name{ScoreMatrixControl}
\alias{ScoreMatrixControl}
\alias{ScoreMatrixControl,ScoreMatrix,ScoreMatrix-method}
\title{Make ScoreMatrixControl from ScoreMatrix objects of IP and control samples}
\usage{
ScoreMatrixControl(IP, control)
}
\arguments{
\item{IP}{\code{\link{ScoreMatrix}} object storing an IP sample}

\item{control}{\code{\link{ScoreMatrix}} object storing a control sample}
}
\value{
\code{ScoreMatrixControl} object
}
\description{
The function creates a \code{ScoreMatrixControl} object that stores a \code{ScoreMatrix} containing a signal values of IP sample as well as a \code{ScoreMatrix} object with control sample. 
This object is an input in \code{\link{enrichmentMatrix}} function that compute enrichment over IgG or input DNA control.
}
