% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrixControl.R
\docType{methods}
\name{ScoreMatrixControl}
\alias{ScoreMatrixControl}
\alias{ScoreMatrixControl,ScoreMatrix,ScoreMatrix-method}
\title{Make ScoreMatrixControl from ScoreMatrix objects of IP and control samples}
\usage{
ScoreMatrixControl(IP, control)
}
\arguments{
\item{IP}{\code{\link{ScoreMatrix}} object storing an IP sample}

\item{control}{\code{\link{ScoreMatrix}} object storing a control sample}
}
\value{
\code{ScoreMatrixControl} object
}
\description{
The function creates a \code{ScoreMatrixControl} object that stores a \code{ScoreMatrix} 
containing a signal values of IP sample as well as a \code{ScoreMatrix} object with control sample. 
This object is an input in \code{\link{enrichmentMatrix}} function that compute enrichment over IgG or input DNA control.
}
\examples{
 
#load IP and control BAM files and create ScoreMatrix objects
source("http://bioconductor.org/biocLite.R")
biocLite('genomationData')
bam.file_IP <- system.file("extdata", 
"wgEncodeBroadHistoneH1hescSuz12051317AlnRep1.chr21.bam", package = "genomationData")
bam.file_c <- system.file("extdata", 
"wgEncodeBroadHistoneH1hescControlStdAlnRep1.chr21.bam", package = "genomation")
data(promoters)
IP <- ScoreMatrix(target = bam.file_IP, windows = promoters, type = 'bam')
control <- ScoreMatrix(target = bam.file_c, windows = promoters, type = 'bam')

# create a ScoreMatrixControl object
ScoreMatrixControl(IP, control)

}
\seealso{
\code{\link{ScoreMatrixListControl}}, \code{\link{ScoreMatrix}}, \code{\link{enrichmentMatrix}}
}
