% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomation-classes.R
\docType{class}
\name{ScoreMatrixList-class}
\alias{ScoreMatrixList-class}
\title{An S4 class for storing a set of \code{ScoreMatrixList}}
\description{
The resulting object is an extension of a \code{list} object, where each element corresponds to a score matrix object
}
\section{Slots}{

\describe{
\item{\code{name}}{A character vector containing names of all stored \code{ScoreMatrix} objects.}
}}

\section{Constructors}{

see \code{\link{ScoreMatrixList}}
}

\section{Coercion}{

as(from, "ScoreMatrixList"): Creates a \code{ScoreMatrixList} object from a list containing \code{\link{ScoreMatrix}} or \code{\link{ScoreMatrixBin}} objects.
}

\section{Subsetting}{

In the code snippets below, x is a ScoreMatrixList object.

\code{x[[i]]},\code{x[[i]]}: Get or set elements \code{i}, where \code{i} is a numeric or character vector of length 1.

\code{x$name}, \code{x$name}: value: Get or set element \code{name}, where \code{name} is a name or character vector of length 1.
}

\seealso{
\code{\link{ScoreMatrixList}}
}
