\name{scoreMatrixList-class}
\alias{scoreMatrixList-class}
\title{An S4 class for storing a set of \code{scoreMatrixList}}
\description{
  The resulting object is an extension of a \code{list}
  object, where each element corresponds to a score matrix
  object
}
\seealso{
  \code{\link{scoreMatrixList-methods}}
}

