% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ScoreMatrixList-class}
\alias{ScoreMatrixList-class}
\title{An S4 class for storing a set of \code{ScoreMatrixList}}
\description{
The resulting object is an extension of a \code{list} object, where each element corresponds to a score matrix object
}
\seealso{
\code{\link{ScoreMatrixList}}
}

