\docType{methods}
\name{scoreMatrixList}
\alias{scoreMatrixList}
\title{Construct a list of scoreMatrixObjects that can be used for plotting}
\usage{
  scoreMatrixList(...)
}
\arguments{
  \item{...}{corresponds to 1 or more scoreMatrix objects}
}
\value{
  returns a \code{scoreMatrixList} object
}
\description{
  Construct a list of scoreMatrixObjects that can be used
  for plotting
}

