% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMatrixListControl.R
\docType{methods}
\name{ScoreMatrixListControl}
\alias{ScoreMatrixListControl}
\alias{ScoreMatrixListControl,ScoreMatrixList,ScoreMatrixList-method}
\title{Make ScoreMatrixListControl from ScoreMatrixList objects of IP and control samples}
\usage{
ScoreMatrixListControl(IP, control)
}
\arguments{
\item{IP}{\code{\link{ScoreMatrixList}} object storing IP samples}

\item{control}{\code{\link{ScoreMatrixList}} storing control samples}
}
\value{
\code{ScoreMatrixListControl} object
}
\description{
The function creates a \code{ScoreMatrixListControl} object that stores \code{ScoreMatrixControl} objects.
This object is an input in \code{\link{enrichmentMatrix}} function that compute enrichment over IgG or input DNA control.
}
