\name{show}
\alias{show}
\alias{show,annotationByFeature-method}
\alias{show,annotationByGenicParts-method}
\alias{show,randomEnrichment-method}
\title{show method for some of the genomation classes}
\arguments{
  \item{object}{object of class randomEnrichment}
}
\description{
  show method for some of the genomation classes
}

