\name{show,randomEnrichment-method}
\alias{show,ScoreMatrix-method}
\alias{show,ScoreMatrixList-method}
\alias{show,annotationByFeature-method}
\alias{show,annotationByGenicParts-method}
\alias{show,randomEnrichment-method}
\title{show method for some of the genomation classes}
\usage{
  \S4method{show}{randomEnrichment}(object)

  \S4method{show}{annotationByGenicParts}(object)

  \S4method{show}{annotationByFeature}(object)

  \S4method{show}{ScoreMatrix}(object)

  \S4method{show}{ScoreMatrixList}(object)
}
\arguments{
  \item{object}{object of class randomEnrichment}
}
\description{
  show method for some of the genomation classes
}

