% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomation-classes.R, R/readAnnotate.R, R/scoreMatrix.R, R/scoreMatrixList.R
\docType{methods}
\name{show,RandomEnrichment-method}
\alias{show,AnnotationByFeature-method}
\alias{show,AnnotationByGeneParts-method}
\alias{show,RandomEnrichment-method}
\alias{show,ScoreMatrix-method}
\alias{show,ScoreMatrixList-method}
\title{show method for some of the genomation classes}
\usage{
\S4method{show}{RandomEnrichment}(object)

\S4method{show}{AnnotationByGeneParts}(object)

\S4method{show}{AnnotationByFeature}(object)

\S4method{show}{ScoreMatrix}(object)

\S4method{show}{ScoreMatrixList}(object)
}
\arguments{
\item{object}{object of class RandomEnrichment}
}
\value{
Shows the dimension of the ScoreMatrix

Shows the number of matrices and their sizes
}
\description{
show method for some of the genomation classes
}

