\docType{methods}
\name{show,RandomEnrichment-method}
\alias{show,AnnotationByFeature-method}
\alias{show,AnnotationByGeneParts-method}
\alias{show,RandomEnrichment-method}
\alias{show,ScoreMatrix-method}
\alias{show,ScoreMatrixList-method}
\title{show method for some of the genomation classes}
\usage{
\S4method{show}{RandomEnrichment}(object)

\S4method{show}{AnnotationByGeneParts}(object)

\S4method{show}{AnnotationByFeature}(object)

\S4method{show}{ScoreMatrix}(object)

\S4method{show}{ScoreMatrixList}(object)
}
\arguments{
  \item{object}{object of class RandomEnrichment}
}
\description{
  show method for some of the genomation classes
}

