% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.R
\docType{methods}
\name{[,ScoreMatrix,ANY,ANY,ANY-method}
\alias{[,ScoreMatrix,ANY,ANY,ANY-method}
\alias{[,ScoreMatrix-method}
\title{Extract method for a ScoreMatrix object.}
\usage{
\S4method{[}{ScoreMatrix,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{the \code{\link{ScoreMatrix}} object}

\item{i}{numeric value}

\item{j}{numeric value}
}
\description{
Extract method for a ScoreMatrix object.
}
